move entry to fasta generator to shared folder
